__init__.py
from mammoth.utils import globals
from mammoth.models import get_model_names, register_model, ContinualModel
from mammoth.datasets import get_dataset_names, register_dataset
from mammoth.datasets.utils import set_default_from_args
from mammoth.datasets.utils.continual_dataset import ContinualDataset, store_masked_loaders
from mammoth.backbone import get_backbone_names, register_backbone, MammothBackbone, ReturnTypes
from mammoth.utils.notebooks import load_runner, get_avail_args
from mammoth.utils.training import train
from mammoth.utils.conf import base_path, get_device
from mammoth.utils.buffer import Buffer
from mammoth.utils.args import add_rehearsal_args

__all__ = [
    "get_dataset_names",
    "get_model_names",
    "get_backbone_names",
    "load_runner",
    "get_avail_args",
    "train",
    "register_model",
    "register_dataset",
    "register_backbone",
    "ContinualModel",
    "ContinualDataset",
    "MammothBackbone",
    "base_path",
    "get_device",
    "ReturnTypes",
    "Buffer",
    "add_rehearsal_args",
    "globals",
    "store_masked_loaders",
    "set_default_from_args",
]